Feat: Add components, routes folder, Relocate Movie Component and Add Detail Component
